refactor(windows): split up into modules
